listas/alerta_obs/01_alerta_obs/python/requirements.txt
obs-websocket-py
paho-mqtt